Stateful Reactor And Shield
The new stateful reactor build itself based on data provided to it by the reactWith method or the data field.

Sheilf catches errors in the widget tree and builds an error state UI, it is possible to use shield to log and gracefully recover from error states

The old reactor has been renamed to catalyst.

rebuild has been renamed to react